Some minor refactoring of the audio components.
* Renamed Say Effector and Hear Perceptor to Speaker and Microphone respectively.
* Adapted audio network protocol to allow for multiple audio sensors and actuators.
* Added sparker volume to speaker action and action interface.
* Added some docs for the microphone perceptor and speaker effector network protocol.
* Fixed parsing order in action parser (motor actions are not uniquely identifyable and thus should be parsed last).

+++ src/rcsssmj/sim/sensors.py
@@ -56,7 +56,7 @@ class Microphone(CustomSensor):
         self.volumes: NDArray[np.float64] = np.zeros(0, dtype=np.float64)
         """The volumes at which the individual messages arrived."""
 
-        self.origins: NDArray[np.float32] = np.zeros((3, 0), dtype=np.float64)
+        self.origins: NDArray[np.float64] = np.zeros((3, 0), dtype=np.float64)
         """The sensor local origins of the sound transmissions."""
 
         self.distances: NDArray[np.float64] = np.zeros(0, dtype=np.float64)
